Variable redefinitions report correct name
The 'put' method in the interpreter used the identifying name of a
method for both identifying and reporting. This wasn't an issue for
variables, as their identifying and pretty names are the same, but it
meant that an error from creating a method whose name already existed
were not correct.

On the theme of ensuring requests are consistent, this commit also
includes some request tests for def and var.