Parse and interpret direct and indirect returns
The lexer, parser and AST now support return statements. The interpreter
implements returns by storing the callback of a method in the scope
stack, and executes a return by finding the nearest callback in scope
and calling it, passing the return's expression if it has one. It does
this without executing its own callback, effectively jumping from its
current location to the end of the method execution.