Differentiate Grace and JS objects on request
Default JavaScript methods are now not available on Grace objects.
Conversely, standard JavaScript objects will appear to have the default
Grace methods, as well as support for getting and setting properties
that are not functions. The REPL now gets any object to mimic the
behaviour of asString if it's implementation is the same as the default
JavaScript toString.