Support more complicated whitespace parsing
Newlines may now appear in the middle of a separate parse, so long as
they are indented further than the start of the actual line. This
involves updates to the lexer as well as the parser, as consecutive
newlines are now mashed together, and the final line remembers how much
space appears after it.

The parser also keeps track of general indentation, enforcing indents on
newlines following open braces and ensuring that indentation remains
consistent across lines that should be at the same indentation. The
parser is no longer context-free, and though this could be maintained by
performing rewriting in the lexer, it seems more sensible to make the
parser smart if it keeps the lexer dumb.