Parse and interpret inherits clause
Inheritance can now be achieved with an inherits clause, which is
strictly parsed to contain a request. Methods which tail-return an
object have an inherit method that is called instead of the method
itself. The inherit method splits the body and interprets the last node
with a predetermined self value. Future work requires clarifying the
lookup semantics, and whether inherited methods can be requested
unqualified.