Handle late returns correctly
Methods that had returned were leaving their callback behind in the
stack, and late returns in blocks were finding that callback and
reapplying the exit of the method. Now methods set the callback to null
when exiting, so late returns recognise that they are returning from an
already exited method. The previous error case remains, but indicates
that the return was never inside another method.

A few method names have also been changed in the interpreter.