Curry multipart methods
The underlying implementation of methods took lists of arguments to each
part of a method's signature, which was easy to process but meant that
single-part methods were not exposed as regular Javascript functions.
This change identifies arguments for each part by taking the arguments
for one part, and then returning a function that will take the next.
Once all arguments are given, the method runs. Single-part methods now
behave like regular JavaScript functions, as they run the method
straight away.