Literals may appear in requests without parens
Numbers, strings, and blocks may now be used as single arguments to
requests without having to wrap them in parentheses. This works in any
part of a multipart method, and blocks, but not numbers or strings, may
have the request automatically closed by a following method request.

This commit also resolves a bug in the multipart method parameter system
that was exposed by the multipart literal test.