Expand string functionality
String literals are now lexed as a single token, which ensures that
whitespace and escapes are parsed correctly inside of a string. Escape
characters and unicode points are available as well. The REPL now
surrounds strings in quotes when printing them as the result of an
evaluation to emphasise that the value is a string, and escapes control
characters to avoid printing a malformed output.