Lexer is on-demand, now in control of parser
Rather than have the lexer perform all the lexing before passing the
result to the parser, the lexer produces each token on-demand. The
parser controls the lexer, lexing the next token only when needed. This
is necessary to support interpolation, where the lexer needs help to
determine whether a token belongs in an expression or a string literal.
In the future, the lexer could be expanded to stream from a buffer
rather than reading from a whole string.