Expressions can be interpolated into strings
String parsing now takes into account braces in the string, which ends a
string literal and indicates that the encased expression should be
interpolated. The parsing can handle nested braces, as well as changes
in indentation inside the interpolation.

A bug in the AST pretty-printing that did not take into account
operators has also been fixed.