Add type literals and method pattern matching
Methods now check their parameters and return values against the
associated patterns, and those patterns are themselves checked against a
'Pattern' pattern. Type literals are now available, and they construct
runtime pattern objects that check for the presence of each specified
method. They currently only check that methods with the necessary name
exist, but in the future should check parameter length and pattern
compatibility.

Internally, when the interpreter produces curried methods every
intermediate function is now asynchronous, as each part checks another
pattern and that check may also be asynchronous. This involved a
substantial change to the definitions of 'method' and 'request'. A
number of builtin patterns have also been introduced into the runtime,
and it would be nice for these to be implemented as Grace code in the
future.