Prefer definitions closer defined closer than self
Defining a local variable with the same name as a method on self would
always be ignored in favour of the method. Now the search prefers local
variables until it reaches the surrounding object boundary, at which
point it searches on self, and then continues if it is still not found.