Fix incorrect implementation of multipart methods
The existing implementation of multipart methods did not correctly
resolve the arguments and callback. In the new implementation each part
returns the next method to call until the final part is processed.
Single part methods still resemble standard asynchronous JavaScript
functions.

This also resolves an issue with the output of asString for objects with
more than one method.