Allow parameters to be ignored with underscore
Parameters may now be ignored by naming them as an underscore. They may
still be given patterns, which will still be checked. As part of the
change to the interpreter that ignores parameters named as underscores,
the mechanism for checking patterns has been tidied up.

The test harness will now no longer stall when encountering a non-Grace
file in the test directory as well.