Add generic parameters to signatures and requests
Generics parameters now parse on method and class signature parts, and
requests. The request parsing is achieved through proper backtracking,
but peeking lookahead is still preferred. The whole lookahead system
should probably be moved into the lexer at some point.

Dynamically, generics are evaluated similarly to regular arguments, and
they must match the Pattern pattern. They are passed in to a method
either as a 'generics' property on each part argument list or, in the
case of single-part methods, on the callback. Parts which have generic
parameters currently no longer evaluate their patterns and cannot check
them dynamically, because it's not well-defined how to evaluate the
patterns in the presence of a generic parameter.