Add named type declarations, with generics
Types can now be named with declarations and optional generics. The
current dynamic semantics aren't well defined so at the moment generic
arguments are ignored to avoid infinite recursion in mutually recursive
types. Generics may be ignored, but if supplied they must have the
appropriate number of arguments.

Generic request parsing is also updated to allow arbitrary expressions,
which permits nested generics. This involves a little hack in the parser
to split tokens such as '>>' into the their separate components.