Add annotations to objects and declarations
Objects and declarations can now be annotated with the 'is' keyword. The
annotations are currently evaluated and then thrown away, as it's not
well defined what an annotation actually is. Parsing of annotations is
stricter than normal expressions, as it must not consume the method and
class bodies or var assignment operators.