Pass sane JSLint rules
All of the JavaScript has been updated to pass a large subset of the
JSLint rules, with the exception of the runtime (which must temporarily
fail the mutual dependencies and 'new' tests tests) and the unicode
module (which is expected to be removed for a standard Node library in
the future). This change includes the scripts and test harness.

All future changes are expected to continue to pass JSLint runs.