Add dialect scoping and interactive modules
Dialects are now parsed and interpreted, loading the requested module
and replacing the default prelude with the resulting object.

The main script now accepts an option to indicate interactive mode that
can be combined with a file name to load that module as the dialect of
the REPL. A new 'enter' method on the interpreter allows the creation of
a consistent object as the self value for the interactive mode.