Always use pretty output for methods and requests
The existing mechanism for building the name of a method or request was
frustrating because it required rebuilding the name with a different
parameter when the pretty form was required. The short form can always
be generated by stripping off the parentheses from a pretty form, so
the pretty form is now the only output of the name method.