Add exception and packet primitives
The exception and exception packet primitives have been added as
constructors to the primitive definitions, and the internal definitions
have been updated to use them. The parser, interpreter and runtime
machinery still throw strings, but these are now reported as internal
errors. They should be updated to use the new exceptions in the future.

The packets do not yet build up a stack trace, and native errors are not
wrapped, both of which should also be patched. The InternalError
exception will catch native errors though.