Add stack traces to exception packets
Exception packets now trace the methods and objects they were in and
dump this information if they are not caught. The trace includes the
name of methods, the object the method was called on, and the module the
method was defined in. This information is not yet exposed as through
methods on the packets themselves.