Add outer, boolean keywords, common indentation
The use of true and false are now only allowed as direct requests, and
requesting outer will now return the self value above the current one in
scope. Chains of outer resolve to a single AST node that resolve to the
appropriate self value in scope.

The indentation of the tests has been standardised to two spaces.