Add method getter and arguments helper to apply
Specialised uses of apply had to reference the receiver of the method
and then separately retrieve the method from the receiver. If the method
took no arguments, apply still had to take a nested array. Now apply
optionally accepts a string instead of a method, in which case it
retrieves the method with that name from the receiver, and the arguments
default to empty.