Parse type literals at start of line as expression
Type literals that began at the start of a line are confused with type
declarations, as they start with the same keyword. The parser for type
declarations had a case to deal with this possibility, but it only
parsed the literal and did not continue on the line. The parser will now
parse a full expression, included method requests, binary operators, and
a new line.