Rename class to constructor, add dotted class
In an effort to come back into line with the language design, and given
that the existing 'classes' were not really classes, the 'dotted class'
syntax has been added to the language as sugar for a single-method
object. The old class syntax is now available with the 'constructor'
keyword, and the tests have been updated accordingly.