Fold comments into preceding newlines
Comments broke up the new line folding and produced more than one new
line at a time. Now the newline folding has been tidied up and can
account for comments. The leading character of a comment is no longer
temporarily lexed as a symbol when the second character is present.