Adjust signature lookahead to include operators
The lookahead on signatures only tested for identifiers, which prevented
operators from appearing in types. Now the lookahead will test for
either an identifier or an operator. The branching for parsing various
types of signatures has also been tidied up, and a test for an
assignment method has been added.