Add at() and size methods to string
There was no way to retrieve characters or the size of a string without
using the primitive interface. The at() and size methods almost brings
strings in line with the list type sans doForEach(), but this will soon
be moved into the prelude. A hidden bug where a block was returning
undefined instead of done in the list implementation has also been
resolved.