Allow newlines before a close parenthesis
The parser wasn't permitting a new line to follow an expression in
parenthesis when dropping back to a previous indentation. Now an indent
in parentheses are just treated as a continutation line, but a newline
is permitted at the end.

The error message "Unexpected use of" has been changed to "Unexpected
appearance of" to reflect the fact that some tokens aren't actually
written by the programmer.