Add class to AST and hoist appropriately
As the creation of a class object is side-effect free, it is safe to
evaluate it immediately, allowing it to be used prior to its definition
in the same way as methods. As such, classes are now a separate entity
in the AST rather than being parsed as sugar, and have a similar
interface to methods. The interpreter code for methods has been expanded
slightly to handle the different meaning of the class body.

Signatures are now also represented as an AST node rather than as a
native list, allowing them to generate their own name and toString
representations.