Hoist methods and variables before types
Because type declarations were hoisted before methods and variables, use
of methods and variables in types produced an UnresolvedRequest despite
the definitions being in scope. Now methods and variables are hoisted
first, which is side-effect free, and types may use methods so long as
those methods do not use the type (otherwise an UninstantiatedType error
occurs). Using variables in types is still guaranteed to produce an
error, but it will now produce the more sensible UndefinedValue rather
than UnresolvedRequest.

Note that variables are not hoisted in a non-object body, so this
remains a problem there. It's not clear if there is a solution to this
problem other than not hoisting type declarations in that circumstance.