Correct undefined method return internal error
The internal error that resulted from a method returning either
undefined or null logged to the console and then produced an
unrenderable exception. This change removes the log and correctly
converts the message to a Grace string before raising the error.

An overly long line in the interpreter has also been trimmed.