Improve stack trace reports for new evaluations
Some new features such as annotations were not being logged in the stack
trace when exceptions occurred. The node reporting in the put method has
been abstracted out to work at any point, and is used to report problems
with annotations on a given node.

Requests to types with parameters will now report the request for a type
rather than as a method.