Use type aliases names in type asString
Defining a type alias now creates a proxy with a name and uses that
name instead of the structure of the type when using asString. Applying
generics includes those types in the stringification as well. The
built-in types have been updated to include the generic information.

Built-in methods have been split out into a separate module as well, and
a method 'delegateTo' has been added to handle object inheritance.