Convert native objects when calling native code
Rather than having to manually convert builtin objects to their native
representation when invoking JavaScript functions, the lookup function
will generate a wrapper when it detects a JavaScript function that does
this automatically. This is handled by implementing an 'asPrimitive'
function which returns whatever the native representation of the object
is. Note that the existing asPrimitive* functions still exist, and some
builtin objects now implement multiple variants.

This change also tidies up the number type to include the builtin binary
operators.