Adjust variable lookup and method overriding
A full variable lookup will search all scopes and every self value in
scope if a definition cannot be found in the immediate scope or self
value, and an ambiguity error occurs if multiple definitions are found.

Variables are no longer permitted to override methods in a super-object.
This avoids tricky situations where code that relies on variables having
values can be broken by an overriding variable.