Tidy up builtin boolean and list definitions
The andAlso() orElse() method has been relegated to short circuiting
booleans only, and replaced by the ifTrue() ifFalse() method and
companions ifTrue() and ifFalse(). Due to the issue with infinite
recurion on assertions, both ifTrue() ifFalse() and orElse() are both
untyped.

The doForEach() method on lists has been renamed to just do(). A 'Do'
type has been added to the prelude and the method for() do() has been
implemented as an associated control structure.