Convert primitive values for getters and setters
Grace objects were converted back and forth from primitive
representations when calling out to JavaScript code, but only in
functions. Now getters and setters on JavaScript objects will also
perform a conversion before setting a value and after getting a value.

This commit also includes a minor change to the types that renames the
erroneously named 'Any' type to 'Object'.