Add names to invalid type error details
When multiple methods with the same name appeared in the same type, the
error message did not include the name of the duplicate method or the
name of the type if it appeared in an alias. Now alias names are
included in error stack traces and the InvalidType error are raised with
the name of the method which appears more than once.