Add dialect checking mechanism with AST
Modules which define a method named checked will now be passed the AST
nodes of the modules which use them as a dialect. A checker method may
raise a Checker Failure exception with a node from the AST to indicate a
static error at that point. The AST has been adjusted to be make
processing a checker method a little more sensible.

Some bugs in the JavaScript to Grace method lookup have been identified
and resolved by this change. The lookup will now also pretend that
JavaScript arrays have the built-in List methods.