Rename list to sequence, complete coercer
The list objects stringified to as a 'sequence', and have now been
renamed throughout to use that name instead of list. A sequence class
is now available in the prelude. The type 'Do' has been moved from the
Grace prelude into the runtime, as it is now used internally.

The lookup mechanism on JavaScript objects has been updated to handle
all primitive types, providing methods as though they were the built-in
Grace objects when no such property exists on the primitive value.

This change also tidies up a few misnames in the primitive definitions,
and the built-in Boolean type.