Simplify and specialise default object stringify
Rendering the names of every method in an object by default ballooned
too easily. Now Grace objects simply stringify to "object", but the
interpreter will add more information to object's asString methods if
they do not update the default, including the class and constructor name
of the object or the surrounding module name if no such method can be
found. Classes now stringify to their name.