Tidy up request tracing on raised exception
Requests were added to the stack trace when evaluating one of their
arguments raised an exception, but the argument will already report
itself, so reporting the request afterwards is not useful. This also
tidies up the post-request error reporting, tying in inheritance to
ensure that requests in inherits clauses are reported correctly.